add pinpoint resources

+++ src/integ_test_resources/ios/sdk/integration/cdk/cdk_integration_tests_ios/file_utils.py
@@ -1,16 +1,26 @@
-import re
+import re, os
 
 def replace_in_file(file_path: str, pattern: str, substitute: str):
     """
     Find the pattern in the file and substitute it.
     """
-    # Read contents from file as a single string
-    file_handle = open(file_path, 'r')
-    file_string = file_handle.read()
-    file_handle.close()
-    file_string = (re.sub(pattern, substitute, file_string))
+    if os.path.exists(file_path):
+        # Read contents from file as a single string
+        try:
+            file_handle = open(file_path, 'r')
+            file_string = file_handle.read()
+            file_handle.close()
+        except:
+            raise IOError("Unable to read from specified file")
 
-    # Write contents back to the file.
-    file_handle = open(file_path, 'w+')
-    file_handle.write(file_string)
-    file_handle.close()
+        file_string = (re.sub(pattern, substitute, file_string))
+
+        # Write contents back to the file.
+        try:
+            file_handle = open(file_path, 'w+')
+            file_handle.write(file_string)
+            file_handle.close()
+        except:
+            raise IOError("Unable to write to specified file")
+    else:
+        raise FileNotFoundError("Unable to find specified file")
